added new pages
added login, ownerView, and fixed appointementForm. Also seperated the css and javaScript but cannot for ownerView and appointmentForm. Will need to fix that. Also added hardcoded code to see what it would look like. Right now there is a hardcoded user with the email of: test@example.com. With a password of 123. in the ownerView I added 3 test appointments. Still need to further develop and review javascript for ownerView.